Simpler getting-up-and-running
I don't see the need for configuring the `INDEX_NAME` or `DOC_DIR` variables, as they both have defaults we can use.

Setting a custom doc dir also has the imitation that it needs to be part of the docker build context, so anything outside of the project dir would not work. We can either explain this or just tell people to copy over stuff to the existing dir.